donut/requirements.txt
transformers
torch
torchvision
Pillow
sentencepiece
